Refactor example models
