+++ tests/test_models/phelps_deaton.py
@@ -1,55 +1,44 @@
-"""Define example model specifications."""
+"""Example specifications of the Phelps-Deaton model."""
 
 import jax.numpy as jnp
 
+# ======================================================================================
+# Numerical parameters and constants
+# ======================================================================================
+
+N_GRID_POINTS = {
+    "states": 100,
+    "choices": 200,
+}
+
 RETIREMENT_AGE = 65
-N_CHOICE_GRID_POINTS = 500
-N_STATE_GRID_POINTS = 100
+
+# ======================================================================================
+# Model functions
+# ======================================================================================
 
 
-def phelps_deaton_utility_with_shock(
-    consumption,
-    working,
-    delta,
-    additive_utility_shock,
-):
-    return jnp.log(consumption) + additive_utility_shock - delta * working
-
-
-def phelps_deaton_utility(consumption, working, delta):
+# --------------------------------------------------------------------------------------
+# Utility functions
+# --------------------------------------------------------------------------------------
+def utility(consumption, working, delta):
     return jnp.log(consumption) - delta * working
 
 
-def phelps_deaton_utility_with_filter(
+def utility_with_filter(
     consumption,
     working,
     delta,
     lagged_retirement,  # noqa: ARG001
 ):
-    return jnp.log(consumption) - delta * working
+    return utility(consumption=consumption, working=working, delta=delta)
 
 
+# --------------------------------------------------------------------------------------
+# Auxiliary variables
+# --------------------------------------------------------------------------------------
 def working(retirement):
     return 1 - retirement
-
-
-def next_wealth_with_shock(
-    wealth,
-    consumption,
-    working,
-    wage,
-    wage_shock,
-    interest_rate,
-):
-    return interest_rate * (wealth - consumption) + wage * wage_shock * working
-
-
-def next_wealth(wealth, consumption, working, wage, interest_rate):
-    return (1 + interest_rate) * (wealth - consumption) + wage * working
-
-
-def consumption_constraint(consumption, wealth):
-    return consumption <= wealth
 
 
 def wage(age):
@@ -60,6 +49,23 @@ def age(_period):
     return _period + 18
 
 
+# --------------------------------------------------------------------------------------
+# State transitions
+# --------------------------------------------------------------------------------------
+def next_wealth(wealth, consumption, working, wage, interest_rate):
+    return (1 + interest_rate) * (wealth - consumption) + wage * working
+
+
+# --------------------------------------------------------------------------------------
+# Constraints
+# --------------------------------------------------------------------------------------
+def consumption_constraint(consumption, wealth):
+    return consumption <= wealth
+
+
+# --------------------------------------------------------------------------------------
+# Filters
+# --------------------------------------------------------------------------------------
 def mandatory_retirement_filter(retirement, age):
     return jnp.logical_or(retirement == 1, age < RETIREMENT_AGE)
 
@@ -68,9 +74,13 @@ def absorbing_retirement_filter(retirement, lagged_retirement):
     return jnp.logical_or(retirement == 1, lagged_retirement == 0)
 
 
+# ======================================================================================
+# Model specification and parameters
+# ======================================================================================
+
 PHELPS_DEATON = {
     "functions": {
-        "utility": phelps_deaton_utility,
+        "utility": utility,
         "next_wealth": next_wealth,
         "consumption_constraint": consumption_constraint,
         "working": working,
@@ -83,7 +93,7 @@ PHELPS_DEATON = {
             "grid_type": "linspace",
             "start": 0,
             "stop": 100,
-            "n_points": N_CHOICE_GRID_POINTS,
+            "n_points": N_GRID_POINTS["choices"],
         },
     },
     "states": {
@@ -91,7 +101,7 @@ PHELPS_DEATON = {
             "grid_type": "linspace",
             "start": 0,
             "stop": 100,
-            "n_points": N_STATE_GRID_POINTS,
+            "n_points": N_GRID_POINTS["states"],
         },
     },
     "n_periods": 3,
@@ -100,7 +110,7 @@ PHELPS_DEATON = {
 
 PHELPS_DEATON_FULLY_DISCRETE = {
     "functions": {
-        "utility": phelps_deaton_utility,
+        "utility": utility,
         "next_wealth": next_wealth,
         "consumption_constraint": consumption_constraint,
         "working": working,
@@ -114,33 +124,16 @@ PHELPS_DEATON_FULLY_DISCRETE = {
             "grid_type": "linspace",
             "start": 0,
             "stop": 100,
-            "n_points": N_STATE_GRID_POINTS,
+            "n_points": N_GRID_POINTS["states"],
         },
     },
     "n_periods": 3,
 }
 
 
-PHELPS_DEATON_WITH_SHOCKS = {
-    **PHELPS_DEATON,
-    "functions": {
-        "utility": phelps_deaton_utility_with_shock,
-        "next_wealth": next_wealth_with_shock,
-        "consumption_constraint": consumption_constraint,
-        "working": working,
-    },
-    "shocks": {
-        "wage_shock": "lognormal",
-        # special name to signal that this shock can be set to zero to calculate
-        # expected utility
-        "additive_utility_shock": "extreme_value",
-    },
-}
-
-
 PHELPS_DEATON_WITH_FILTERS = {
     "functions": {
-        "utility": phelps_deaton_utility_with_filter,
+        "utility": utility_with_filter,
         "next_wealth": next_wealth,
         "consumption_constraint": consumption_constraint,
         "working": working,
@@ -153,7 +146,7 @@ PHELPS_DEATON_WITH_FILTERS = {
             "grid_type": "linspace",
             "start": 1,
             "stop": 100,
-            "n_points": N_CHOICE_GRID_POINTS,
+            "n_points": N_GRID_POINTS["choices"],
         },
     },
     "states": {
@@ -161,7 +154,7 @@ PHELPS_DEATON_WITH_FILTERS = {
             "grid_type": "linspace",
             "start": 0,
             "stop": 100,
-            "n_points": N_STATE_GRID_POINTS,
+            "n_points": N_GRID_POINTS["states"],
         },
         "lagged_retirement": {"options": [0, 1]},
     },
