+++ tests/test_state_space.py
@@ -2,7 +2,6 @@ import jax.numpy as jnp
 import numpy as np
 import pandas as pd
 import pytest
-from lcm.example_models.basic_example_models import PHELPS_DEATON_WITH_FILTERS
 from lcm.interfaces import Model
 from lcm.process_model import process_model
 from lcm.state_space import (
@@ -13,6 +12,8 @@ from lcm.state_space import (
     create_state_choice_space,
 )
 from numpy.testing import assert_array_almost_equal as aaae
+
+from tests.test_models.phelps_deaton import PHELPS_DEATON_WITH_FILTERS
 
 
 def test_create_state_choice_space():
